requirements.txt
flask==2.3.3
opencv-python==4.8.1.78
pillow==10.0.1
pytesseract==0.3.10
numpy==1.24.3